api_server/requirements.txt
aniso8601==8.0.0
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
google-api-core==1.20.0
google-auth==1.17.1
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httpie==2.1.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pytz==2020.1
requests==2.23.0
rsa==4.2
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
google-cloud-pubsub==1.6.0
